backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
psycopg2-binary
Pillow
djangorestframework-simplejwt
celery  
redis
google-generativeai